+++ src/core/Settings.hpp
@@ -45,6 +45,7 @@ struct Settings {
     HotkeyBindingMap hotkeyBindings{
         DefaultHotkeyBindings()};
 
+    bool pinyinSearch{true};
     bool wildcardMatching{false};
     bool numericQuickLaunch{false};
     std::string numericQuickLaunchOrder{"one-to-zero"};
@@ -82,7 +83,8 @@ public:
         bool wildcardMatching,
         bool numericQuickLaunch,
         std::string numericQuickLaunchOrder,
-        bool executeSingleResultImmediately);
+        bool executeSingleResultImmediately,
+        bool pinyinSearch);
     bool SetProviderEnabled(
         std::string id,
         bool enabled);